Correct watchdog interface use for SCU add workaround for IRQ
This commit corrects the issues in the REVERT_ME patches with
the upstream intel-mid watchdog driver when being backported
to the 3.10.x kernel. Also, we've found that the registration
of the IRQ is causing a conflict with the I2C bus. For now we
will disable this in the driver as it will not interfer with
normal operations of the watchdog (the IRQ handler should never
be reached anyway and is only used to print a kernel message).
This will be fixed by a subsequent patch.

Change-Id: I06757bdf66615ce5cf63396269d8ee889a54eed3
Signed-off-by: Brian Wood <brian.j.wood@intel.com>
Reviewed-on: https://android.intel.com/424751